♻️ 스터디룸 마감 d-day,지원자수, 게시물 조회수 integer 타입으로 변경

+++ src/main/java/gaji/service/domain/room/web/dto/RoomResponseDto.java
@@ -38,8 +38,8 @@ public class RoomResponseDto {
         private LocalDate endDay;
         private LocalDate recruitStartDay;
         private LocalDate recruitEndDay;
-        private Long daysLeftForRecruit;
-        private Long applicantCount;
+        private Integer daysLeftForRecruit;
+        private Integer applicantCount;
     }
     @Builder
     @Getter
